+++ internal/service/base/data_source_role.go
@@ -6,25 +6,39 @@ import (
 	"net/http"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
+	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
+	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/patrickcping/pingone-go-sdk-v2/management"
 	"github.com/pingidentity/terraform-provider-pingone/internal/framework"
 	"github.com/pingidentity/terraform-provider-pingone/internal/sdk"
 	"github.com/pingidentity/terraform-provider-pingone/internal/utils"
+	"github.com/pingidentity/terraform-provider-pingone/internal/verify"
 )
 
 // Types
 type RoleDataSource serviceClientType
 
 type RoleDataSourceModel struct {
-	Name        types.String `tfsdk:"name"`
-	Description types.String `tfsdk:"description"`
-	Id          types.String `tfsdk:"id"`
+	Id           types.String `tfsdk:"id"`
+	RoleId       types.String `tfsdk:"role_id"`
+	Name         types.String `tfsdk:"name"`
+	Description  types.String `tfsdk:"description"`
+	ApplicableTo types.Set    `tfsdk:"applicable_to"`
+	Permissions  types.Set    `tfsdk:"permissions"`
 }
+
+var (
+	rolePermissionTFObjectTypes = map[string]attr.Type{
+		"id":          types.StringType,
+		"classifier":  types.StringType,
+		"description": types.StringType,
+	}
+)
 
 // Framework interfaces
 var (
@@ -46,9 +60,17 @@ func (r *RoleDataSource) Schema(ctx context.Context, req datasource.SchemaReques
 
 	const minAttrLength = 1
 
+	roleIdDescription := framework.SchemaAttributeDescriptionFromMarkdown(
+		"The ID of the role to retrieve.  Must be a valid PingOne resource ID.",
+	).ExactlyOneOf([]string{"name", "role_id"})
+
 	nameDescription := framework.SchemaAttributeDescriptionFromMarkdown(
 		"The name of the role to look up.",
-	).AllowedValuesEnum(management.AllowedEnumRoleNameEnumValues)
+	).AllowedValuesEnum(management.AllowedEnumRoleNameEnumValues).ExactlyOneOf([]string{"name", "role_id"})
+
+	applicableToDescription := framework.SchemaAttributeDescriptionFromMarkdown(
+		"A set of strings that specifies the applicable scopes that the role can be assigned to.",
+	).AllowedValuesEnum(management.AllowedEnumRoleAssignmentScopeTypeEnumValues)
 
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
@@ -57,11 +79,22 @@ func (r *RoleDataSource) Schema(ctx context.Context, req datasource.SchemaReques
 		Attributes: map[string]schema.Attribute{
 			"id": framework.Attr_ID(),
 
+			"role_id": schema.StringAttribute{
+				Description:         roleIdDescription.Description,
+				MarkdownDescription: roleIdDescription.MarkdownDescription,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.ExactlyOneOf(path.MatchRelative().AtParent().AtName("name")),
+					verify.P1ResourceIDValidator(),
+				},
+			},
+
 			"name": schema.StringAttribute{
 				MarkdownDescription: nameDescription.MarkdownDescription,
 				Description:         nameDescription.Description,
-				Required:            true,
+				Optional:            true,
 				Validators: []validator.String{
+					stringvalidator.ExactlyOneOf(path.MatchRelative().AtParent().AtName("role_id")),
 					stringvalidator.OneOf(utils.EnumSliceToStringSlice(management.AllowedEnumRoleNameEnumValues)...),
 				},
 			},
@@ -69,6 +102,38 @@ func (r *RoleDataSource) Schema(ctx context.Context, req datasource.SchemaReques
 			"description": schema.StringAttribute{
 				Description: framework.SchemaAttributeDescriptionFromMarkdown("The description of the role.").Description,
 				Computed:    true,
+			},
+
+			"applicable_to": schema.SetAttribute{
+				Description:         applicableToDescription.Description,
+				MarkdownDescription: applicableToDescription.MarkdownDescription,
+				Computed:            true,
+
+				ElementType: types.StringType,
+			},
+
+			"permissions": schema.SetNestedAttribute{
+				Description: framework.SchemaAttributeDescriptionFromMarkdown("A set of strings that represent permissions that have been assigned to the role.").Description,
+				Computed:    true,
+
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"id": schema.StringAttribute{
+							Description: framework.SchemaAttributeDescriptionFromMarkdown("A string that specifies the ID of the permission.").Description,
+							Computed:    true,
+						},
+
+						"classifier": schema.StringAttribute{
+							Description: framework.SchemaAttributeDescriptionFromMarkdown("A string that specifies the resource for which the permission is applicable.").Description,
+							Computed:    true,
+						},
+
+						"description": schema.StringAttribute{
+							Description: framework.SchemaAttributeDescriptionFromMarkdown("A string that specifies the description of the permission and what the permission enables.").Description,
+							Computed:    true,
+						},
+					},
+				},
 			},
 		},
 	}
@@ -116,49 +181,77 @@ func (r *RoleDataSource) Read(ctx context.Context, req datasource.ReadRequest, r
 		return
 	}
 
-	var role management.Role
+	var role *management.Role
 
-	// Run the API call
-	var entityArray *management.EntityArray
-	resp.Diagnostics.Append(framework.ParseResponse(
-		ctx,
+	if !data.Name.IsNull() {
 
-		func() (any, *http.Response, error) {
-			return r.Client.ManagementAPIClient.RolesApi.ReadAllRoles(ctx).Execute()
-		},
-		"ReadAllRoles",
-		framework.DefaultCustomError,
-		sdk.DefaultCreateReadRetryable,
-		&entityArray,
-	)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
+		// Run the API call
+		var entityArray *management.EntityArray
+		resp.Diagnostics.Append(framework.ParseResponse(
+			ctx,
 
-	if roles, ok := entityArray.Embedded.GetRolesOk(); ok {
-
-		found := false
-		for _, roleItem := range roles {
-
-			if string(roleItem.GetName()) == data.Name.ValueString() {
-				role = roleItem
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			resp.Diagnostics.AddError(
-				"Cannot find role from name",
-				fmt.Sprintf("The role %s cannot be found in the tenant", data.Name.String()),
-			)
+			func() (any, *http.Response, error) {
+				return r.Client.ManagementAPIClient.RolesApi.ReadAllRoles(ctx).Execute()
+			},
+			"ReadAllRoles",
+			framework.DefaultCustomError,
+			sdk.DefaultCreateReadRetryable,
+			&entityArray,
+		)...)
+		if resp.Diagnostics.HasError() {
 			return
 		}
 
+		if roles, ok := entityArray.Embedded.GetRolesOk(); ok {
+
+			found := false
+			for _, roleItem := range roles {
+
+				if string(roleItem.GetName()) == data.Name.ValueString() {
+					roleItem := roleItem
+					role = &roleItem
+					found = true
+					break
+				}
+			}
+
+			if !found {
+				resp.Diagnostics.AddError(
+					"Cannot find role from name",
+					fmt.Sprintf("The role %s cannot be found in the tenant", data.Name.String()),
+				)
+				return
+			}
+
+		}
+	} else if !data.RoleId.IsNull() {
+
+		// Run the API call
+		resp.Diagnostics.Append(framework.ParseResponse(
+			ctx,
+
+			func() (any, *http.Response, error) {
+				return r.Client.ManagementAPIClient.RolesApi.ReadOneRole(ctx, data.RoleId.ValueString()).Execute()
+			},
+			"ReadOneRole",
+			framework.DefaultCustomError,
+			retryEnvironmentDefault,
+			&role,
+		)...)
+		if resp.Diagnostics.HasError() {
+			return
+		}
+
+	} else {
+		resp.Diagnostics.AddError(
+			"Missing parameter",
+			"Cannot find the requested role. role_id or name must be set.",
+		)
+		return
 	}
 
 	// Save updated data into Terraform state
-	resp.Diagnostics.Append(data.toState(&role)...)
+	resp.Diagnostics.Append(data.toState(role)...)
 	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
 }
 
@@ -177,6 +270,41 @@ func (p *RoleDataSourceModel) toState(v *management.Role) diag.Diagnostics {
 	p.Id = framework.StringOkToTF(v.GetIdOk())
 	p.Name = framework.EnumOkToTF(v.GetNameOk())
 	p.Description = framework.StringOkToTF(v.GetDescriptionOk())
+	p.ApplicableTo = framework.EnumSetOkToTF(v.GetApplicableToOk())
+
+	permissions, d := toStateRolePermissions(v.GetPermissions())
+	diags.Append(d...)
+	p.Permissions = permissions
 
 	return diags
+}
+
+func toStateRolePermissions(rolePermission []management.RolePermissionsInner) (types.Set, diag.Diagnostics) {
+	var diags diag.Diagnostics
+	tfObjType := types.ObjectType{AttrTypes: rolePermissionTFObjectTypes}
+
+	if len(rolePermission) == 0 {
+		return types.SetValueMust(tfObjType, []attr.Value{}), diags
+	}
+
+	flattenedList := []attr.Value{}
+	for _, v := range rolePermission {
+
+		service := map[string]attr.Value{
+			"id":          framework.StringOkToTF(v.GetIdOk()),
+			"classifier":  framework.StringOkToTF(v.GetClassifierOk()),
+			"description": framework.StringOkToTF(v.GetDescriptionOk()),
+		}
+
+		flattenedObj, d := types.ObjectValue(rolePermissionTFObjectTypes, service)
+		diags.Append(d...)
+
+		flattenedList = append(flattenedList, flattenedObj)
+	}
+
+	returnVar, d := types.SetValue(tfObjType, flattenedList)
+	diags.Append(d...)
+
+	return returnVar, diags
+
 }