+++ internal/service/base/service.go
@@ -43,9 +43,11 @@ func DataSources() []func() datasource.DataSource {
 		NewAgreementLocalizationDataSource,
 		NewEnvironmentDataSource,
 		NewEnvironmentsDataSource,
+		NewGatewayDataSource,
 		NewOrganizationDataSource,
 		NewPhoneDeliverySettingsListDataSource,
 		NewRoleDataSource,
+		NewRolesDataSource,
 		NewTrustedEmailDomainDataSource,
 		NewUserRoleAssignmentsDataSource,
 	}