+++ CHANGELOG.md
@@ -2,10 +2,30 @@
 
 NOTES:
 
+* Updated documentation examples to remove reference to deprecated parameters/attributes. ([#603](https://github.com/pingidentity/terraform-provider-pingone/issues/603))
 * `data-source/pingone_role`: Migrated to plugin framework. ([#592](https://github.com/pingidentity/terraform-provider-pingone/issues/592))
 * `data-source/pingone_trusted_email_domain`: Corrected documentation descriptions. ([#593](https://github.com/pingidentity/terraform-provider-pingone/issues/593))
 * `resource/pingone_environment`: Deprecated the `default_population` block and `default_population_id` attribute in favour of the new `pingone_population_default` resource. ([#485](https://github.com/pingidentity/terraform-provider-pingone/issues/485))
 * `resource/pingone_environment`: Removed the ability to import the resource including a default population.  Default populations are now managed with the `pingone_population_default` resource. ([#485](https://github.com/pingidentity/terraform-provider-pingone/issues/485))
+* bump `github.com/google/uuid` v1.3.1 => v1.4.0 ([#623](https://github.com/pingidentity/terraform-provider-pingone/issues/623))
+* bump `github.com/hashicorp/terraform-plugin-framework` v1.4.1 => v1.4.2 ([#623](https://github.com/pingidentity/terraform-provider-pingone/issues/623))
+* bump `github.com/patrickcping/pingone-go-sdk-v2/credentials` v0.4.1 => v0.5.0 ([#623](https://github.com/pingidentity/terraform-provider-pingone/issues/623))
+* bump `github.com/patrickcping/pingone-go-sdk-v2/management` v0.30.0 => v0.31.0 ([#623](https://github.com/pingidentity/terraform-provider-pingone/issues/623))
+* bump `github.com/patrickcping/pingone-go-sdk-v2` v0.10.8 => v0.10.9 ([#623](https://github.com/pingidentity/terraform-provider-pingone/issues/623))
+
+FEATURES:
+
+* **New Data Source:** `pingone_application` ([#598](https://github.com/pingidentity/terraform-provider-pingone/issues/598))
+* **New Data Source:** `pingone_gateway` ([#613](https://github.com/pingidentity/terraform-provider-pingone/issues/613))
+* **New Data Source:** `pingone_roles` ([#599](https://github.com/pingidentity/terraform-provider-pingone/issues/599))
+* **New Resource:** `pingone_group_role_assignment` ([#602](https://github.com/pingidentity/terraform-provider-pingone/issues/602))
+
+ENHANCEMENTS:
+
+* `data-source/pingone_environment`: Added `service.tags` computed attribute that signifies whether the selected PingOne environment was created without example/demo configuration in the DaVinci service. ([#620](https://github.com/pingidentity/terraform-provider-pingone/issues/620))
+* `data-source/pingone_role`: Now supports read-only attributes `applicable_to` and `permissions`. ([#599](https://github.com/pingidentity/terraform-provider-pingone/issues/599))
+* `data-source/pingone_role`: Now supports the ability to look up role data by ID. ([#599](https://github.com/pingidentity/terraform-provider-pingone/issues/599))
+* `resource/pingone_environment`: Added `service.tags` parameter allows for a creation of an environment without example/demo configuration in the DaVinci service. ([#620](https://github.com/pingidentity/terraform-provider-pingone/issues/620))
 
 ## 0.22.0 (17 October 2023)